Add memreport interface
Including a modification to the puc-lua source (by way of
lua-src-rs) this change adds a memory report interface.
This is intended to provide a mechanism to track, in detail
where and how memory is being used inside of the lua runtime.
It is not intended to account for all the memory used by lua
(global state is not tracked for example, nor is the api string
cache) but is focused a/b comparisons when executing Lua code.